[Platinum V] Title: 히스토그램, Time: 24 ms, Memory: 8540 KB -BaekjoonHub

+++ 백준/Platinum/1725. 히스토그램/히스토그램.cc
@@ -3,11 +3,11 @@
 using namespace std;
 
 int n;
-long long res;
-long long arr[100001];
-long long seg[4 * 100001][2];
+int res;
+int arr[100001];
+int seg[4 * 100001][2];
 
-long long init(int x, int s, int e) {
+int init(int x, int s, int e) {
 	if (s == e) {
 		seg[x][0] = arr[s];
 		seg[x][1] = s;
@@ -15,8 +15,8 @@ long long init(int x, int s, int e) {
 		return seg[x][0];
 	}
 	int mid = s + (e - s) / 2;
-	long long a = init(x * 2, s, mid);
-	long long b = init(x * 2 + 1, mid + 1, e);
+	int a = init(x * 2, s, mid);
+	int b = init(x * 2 + 1, mid + 1, e);
 	if (a < b) {
 		seg[x][0] = a;
 		seg[x][1] = seg[x * 2][1];
@@ -29,12 +29,12 @@ long long init(int x, int s, int e) {
 	}
 }
 
-pair<long long, int> find_min(int node, int s, int e, int x, int y) {
+pair<int, int> find_min(int node, int s, int e, int x, int y) {
 	if (e < x || s > y) return { 1e9, -1 };
 	if (x <= s && e <= y) return { seg[node][0], seg[node][1] };
 	int mid = s + (e - s) / 2;
-	pair<long long, int> a = find_min(node * 2, s, mid, x, y);
-	pair<long long, int> b = find_min(node * 2 + 1, mid + 1, e, x, y);
+	pair<int, int> a = find_min(node * 2, s, mid, x, y);
+	pair<int, int> b = find_min(node * 2 + 1, mid + 1, e, x, y);
 	if (a.first < b.first) {
 		return a;
 	}
@@ -46,7 +46,7 @@ pair<long long, int> find_min(int node, int s, int e, int x, int y) {
 
 
 void func(int s, int e) {
-	pair<long long, int> m = find_min(1, 0, n - 1, s, e);
+	pair<int, int> m = find_min(1, 0, n - 1, s, e);
 	if (m.second == -1) return;
 	res = max(res, ((e - s + 1) * m.first));
 	if(s < m.second)
